kvm: New _with_qmp decorator
The decorator takes the hypervisor and the instance object as passed
to all hotplug related commands and ensures a valid QmpConnection()
object is found in hypervisor's qmp attribute. After the method
returns, the qmp connection is closed. This way if we have nested
hotplug methods there is no need to change their prototype in order
to pass the QmpConnection object and no need to worry about proper
connection handling (e.g. init, connect, close).

Signed-off-by: Dimitris Aragiorgis <dimara@grnet.gr>